+++ docs/extra.js
@@ -7,7 +7,6 @@
 ///    - text: "<img src=\"https://travis-ci.org/retostauffer/Rfoehnix.svg?branch=master\" />"
 ///      href: ""
 $(document).ready(function(){
-
     var target = $("div.developers")
     if ( target.length == 0 ) return;
 
@@ -35,5 +34,18 @@ $(document).ready(function(){
     ///foehnix in action? http://ertel2.uibk.ac.at:8080/ertel/foehndiag.php
 });
 
-
-
+$(document).ready(function(){
+    // Custom action hide/show source code
+    $(".hide-Rinput").on("click", function() {
+        var elem = $(this).children(".sourceCode")
+        if (elem.length > 0) {
+            $.each(elem, function() {
+                if ($(this).css("display") == "none") {
+                    $(this).css("display", "block");
+                } else {
+                    $(this).css("display", "none");
+                }
+            });
+        }
+    });
+});